Replace SemanticBlocks with custom BlockDelimiters
This change will move Standard from attempting to enforce Weirich-style semantic blocks, because there will always be edge-cases we can't cover automatically, given how expressive Ruby is and the fact that certain cases are genuinely ambiguous, such that an automated tool cannot reliably determine the authorial intent of a block.

Instead, the new rule will only enforce that single-line blocks be delimited with `{` and `}`, but otherwise allow users to use whatever multiline format they wish.
