+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## unreleased changes
+
+* Relax multi-line block rules, moving away from enforcing semantic blocks to
+  instead allowing code to adhere to whatever multi-line format the author deems
+  best
+
 ## 0.13.0
 
 * Update rubocop from 1.7.0 to [1.10.0](https://github.com/rubocop-hq/rubocop/releases/tag/v1.10.0) enabling:
@@ -12,7 +18,7 @@
 * Update rubocop from 1.7.0 to [1.8.1](https://github.com/rubocop-hq/rubocop/releases/tag/v1.8.1)
 * Enabled [`Style/SlicingWithRange`](https://github.com/testdouble/standard/issues/175)
 
-## 0.11.0 
+## 0.11.0
 
 * Update rubocop-performance from 1.9.1 to [1.9.2](https://github.com/rubocop-hq/rubocop-performance/releases/tag/v1.9.2)
 * Update rubocop from 1.4.2 to [1.7.0](https://github.com/rubocop-hq/rubocop/releases/tag/v1.7.0)
