+++ LICENSE.txt
@@ -2,7 +2,7 @@ Copyright (c) 2019 Test Double, LLC
 
 Portions of these files Copyright (c) 2012-18 Bozhidar Batsov:
   - config/base.yml
-  - lib/standard/cop/semantic_blocks.rb
+  - lib/standard/cop/block_delimiters.rb
   - test/cop_invoker.rb
 
 Permission is hereby granted, free of charge, to any person obtaining
