+++ lib/standard.rb
@@ -7,7 +7,7 @@ require "standard/cli"
 require "standard/railtie" if defined?(Rails) && defined?(Rails::Railtie)
 
 require "standard/formatter"
-require "standard/cop/semantic_blocks"
+require "standard/cop/block_delimiters"
 
 module Standard
 end
