+++ test/standard_test.rb
@@ -4,7 +4,7 @@ class StandardTest < UnitTest
   def test_loads_stuff
     refute_nil RuboCop
     refute_nil Standard::Cli
-    refute_nil RuboCop::Cop::Standard::SemanticBlocks
+    refute_nil RuboCop::Cop::Standard::BlockDelimiters
     assert_instance_of Gem::Version, ::Standard::VERSION
   end
 end
